backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
replicate==0.15.4
firebase-admin==6.2.0
pillow==10.1.0
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
